Category still needs to be an integer though

+++ android/src/main/java/com/zmxv/RNSound/RNSoundModule.java
@@ -59,7 +59,7 @@ public class RNSoundModule extends ReactContextBaseJavaModule implements AudioMa
     final RNSoundModule module = this;
 
     if (module.category != null) {
-      Double category = null;
+      Integer category = null;
       switch (module.category) {
         case "Playback":
           category = AudioManager.STREAM_MUSIC;
